restructring L4 - moved service files to services dir
